refactor(Task-3): Moved uiRoute routes to another routes file
