+++ src/app/app.module.js
@@ -1,5 +1,6 @@
 import angular from 'angular';
 import uiRouter from '@uirouter/angularjs';
+import routes from './app.routes';
 import helloComponent from './hello/hello';
 import worldComponent from './world/world';
 
@@ -7,20 +8,4 @@ const app = angular.module('app', [uiRouter]);
 
 app.component('hello', helloComponent);
 app.component('world', worldComponent);
-
-app.config(['$stateProvider', function (provider) {
-    const helloState = {
-        name: 'hello',
-        url: '/hello',
-        component: 'hello'
-    }
-
-    const constState = {
-        name: 'world',
-        url: '/world',
-        component: 'world'
-    }
-
-    provider.state(helloState);
-    provider.state(constState);
-}]);
+app.config(routes);
