+++ src/app/app.routes.js
@@ -1,0 +1,18 @@
+const config = ['$stateProvider', function (provider) {
+    const helloState = {
+        name: 'hello',
+        url: '/hello',
+        component: 'hello'
+    }
+
+    const constState = {
+        name: 'world',
+        url: '/world',
+        component: 'world'
+    }
+
+    provider.state(helloState);
+    provider.state(constState);
+}];
+
+export default config;
